requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-environ==0.10.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
Pillow==9.5.0
promise==2.3
pycodestyle==2.10.0
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.15
